+++ picodi/helpers.py
@@ -5,7 +5,7 @@ import contextlib
 from typing import TYPE_CHECKING, Any, ParamSpec, TypeVar, overload
 
 import picodi
-from picodi import GlobalScope, Scope
+from picodi import ManualScope, Scope
 
 if TYPE_CHECKING:
     from collections.abc import AsyncGenerator, Callable, Generator
@@ -116,7 +116,7 @@ class _Lifespan:
 
     @contextlib.contextmanager
     def sync(
-        self, scope_class: type[Scope] | tuple[type[Scope]] = GlobalScope
+        self, scope_class: type[Scope] | tuple[type[Scope]] = ManualScope
     ) -> Generator[None, None, None]:
         picodi.init_dependencies(scope_class)
         try:
@@ -126,7 +126,7 @@ class _Lifespan:
 
     @contextlib.asynccontextmanager
     async def async_(
-        self, scope_class: type[Scope] | tuple[type[Scope]] = GlobalScope
+        self, scope_class: type[Scope] | tuple[type[Scope]] = ManualScope
     ) -> AsyncGenerator[None, None]:
         await picodi.init_dependencies(scope_class)
         try: