+++ CHANGELOG.md
@@ -11,7 +11,8 @@ We follow [Semantic Versions](https://semver.org/).
 - Added ability to pass custom `scope_class` to `init_dependencies` and `shutdown_dependencies`
 - Breaking changes:
   - Removed `ParentCallScope`
-  - Renamed `enter_decorator` and `exit_decorator` to `enter_inject` and ``exit_inject`
+  - Renamed `enter_decorator` and `exit_decorator` to `enter_inject` and `exit_inject`
+  - Renamed `LocalScope` and `GlobalScope` to `AutoScope` and `ManualScope`
 
 ## Version 0.14.0